requirements.txt
pandas
pyarrow
confluent-kafka
google-cloud-storage
faker
google-cloud-bigquery
